fetchin API data to get data into Results componnent
