+++ client/src/App.js
@@ -1,25 +1,14 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React, {useState} from 'react';
 import '../src/scss/styles.scss';
 import SearchBox from './components/SearchBox';
 import Results from './components/Results';
 
 const App = () => {
-  const [state, setState] = useState([]);
-  useEffect(() => {
-    const fetchData = async () =>  {
-      const result = await axios(
-        'https://api.mercadolibre.com/items/MLA919739960'
-      );
-      setState(result.data);
-    };
-    fetchData();
-    console.log(state);
-  }, []);    
+  const [term, setTerm] = useState('');
   return (
     <div className="App">
-      <SearchBox />
-      <Results data={state}/>
+      <SearchBox setTerm={setTerm}/>
+      <Results term={term}/>
 
     </div>
   );
